added option to enable/disable payment methods now removing vietnamese diacritics on receipts by default fixed slide remove

+++ src/js/common/network.js
@@ -187,6 +187,9 @@ App.deleteMod = (number, btn, done) => {
 };
 
 App.deleteSlide = (_id, btn) => {
+  if (!_id) {
+    return btn.parent().parent().remove();
+  }
   App.ajaxDeleting(btn);
   return $.ajax({
     type: 'DELETE',
